+++ x/evm/keeper/querier.go
@@ -197,12 +197,8 @@ func queryExportAccount(ctx sdk.Context, path []string, keeper Keeper) ([]byte, 
 		return nil, err
 	}
 
-	balanceInt := keeper.GetBalance(ctx, addr)
-	balance := sdk.NewIntFromBigInt(balanceInt)
-
 	res := types.GenesisAccount{
 		Address: hexAddress,
-		Balance: balance,
 		Code:    keeper.GetCode(ctx, addr),
 		Storage: storage,
 	}