+++ x/evm/types/expected_keepers.go
@@ -9,6 +9,7 @@ import (
 type AccountKeeper interface {
 	NewAccountWithAddress(ctx sdk.Context, addr sdk.AccAddress) authexported.Account
 	GetAllAccounts(ctx sdk.Context) (accounts []authexported.Account)
+	IterateAccounts(ctx sdk.Context, cb func(account authexported.Account) bool)
 	GetAccount(ctx sdk.Context, addr sdk.AccAddress) authexported.Account
 	SetAccount(ctx sdk.Context, account authexported.Account)
 	RemoveAccount(ctx sdk.Context, account authexported.Account)