+++ x/evm/types/genesis.go
@@ -4,8 +4,6 @@ import (
 	"errors"
 	"fmt"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
 	ethcmn "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 )
@@ -22,9 +20,9 @@ type (
 	// GenesisAccount defines an account to be initialized in the genesis state.
 	// Its main difference between with Geth's GenesisAccount is that it uses a custom
 	// storage type and that it doesn't contain the private key field.
+	// NOTE: balance is omitted as it is imported from the auth account balance.
 	GenesisAccount struct {
 		Address string        `json:"address"`
-		Balance sdk.Int       `json:"balance"`
 		Code    hexutil.Bytes `json:"code,omitempty"`
 		Storage Storage       `json:"storage,omitempty"`
 	}
@@ -34,12 +32,6 @@ type (
 func (ga GenesisAccount) Validate() error {
 	if ga.Address == (ethcmn.Address{}.String()) {
 		return fmt.Errorf("address cannot be the zero address %s", ga.Address)
-	}
-	if ga.Balance.IsNil() {
-		return errors.New("balance cannot be nil")
-	}
-	if ga.Balance.IsNegative() {
-		return errors.New("balance cannot be negative")
 	}
 	if ga.Code != nil && len(ga.Code) == 0 {
 		return errors.New("code bytes cannot be empty")